+++ ChangeLog.markdown
@@ -1,8 +1,14 @@
 Changelog for Imbo
 =====================
 
-Imbo-0.1.0
--------------
+Imbo-0.2.0
+----------
 __N/A__
+
+* Fixed #150: Error model can't be formatted using an image formatter
+
+Imbo-0.1.0
+----------
+__2012-12-30__
 
 * Initial release